+++ lib/http_api.dart
@@ -4,8 +4,6 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
 import 'package:http_parser/http_parser.dart';
 import 'package:meta/meta.dart';
 import 'package:http/http.dart' as http;
@@ -20,14 +18,16 @@ part './src/base_api.dart';
 /// Utils
 part './src/utils/exceptions.dart';
 part './src/utils/api_link.dart';
+part './src/utils/debug_link.dart';
+part './src/utils/cache_key.dart';
 
 /// Links
 part './src/links/headers_mapper_link.dart';
-part './src/links/debug_link.dart';
+part 'src/links/logger_link.dart';
 part './src/links/http_link.dart';
 
 /// Requests
-part './src/requests/api_request.dart';
+part 'src/data/api_request.dart';
 
 /// Data
 part './src/data/data.dart';