+++ lib/src/data/api_response.dart
@@ -1,16 +1,19 @@
 part of http_api;
 
 class ApiResponse extends http.Response {
-  final ObjectId _id;
+  final ApiRequest apiRequest;
 
-  /// Id of current ApiRequest
-  ObjectId get id => _id;
+  /// Id of current [ApiResponse].
+  /// The same as related [ApiRequest] id.
+  ObjectId get id => apiRequest.id;
+
+  /// ApiRequest object creation timestamp.
+  final DateTime createdAt;
 
   /// Here you can assing your data that will be passed to the next link
   final Map<String, dynamic> linkData;
 
   /// Time when [ApiResponse] was created.
-  final DateTime received;
 
   /// Represent reponse success
   ///
@@ -28,10 +31,9 @@ class ApiResponse extends http.Response {
       statusCode == 308;
 
   ApiResponse.fromHttp(
-      ApiRequest apiRequest, http.BaseRequest request, http.Response response)
+      this.apiRequest, http.BaseRequest request, http.Response response)
       : linkData = apiRequest.linkData ?? <String, dynamic>{},
-        received = DateTime.now(),
-        _id = apiRequest.id,
+        createdAt = DateTime.now(),
         super(
           response.body,
           response.statusCode,
@@ -41,4 +43,19 @@ class ApiResponse extends http.Response {
           persistentConnection: response.persistentConnection,
           reasonPhrase: response.reasonPhrase,
         );
+
+  String toJson() {
+    return jsonEncode(<String, dynamic>{
+      "id": id.hexString,
+      if (apiRequest.key != null) "key": apiRequest.key.value,
+      "createdAt": createdAt.toIso8601String(),
+      "redirect": redirect,
+      "ok": ok,
+      "headers": headers,
+      "body": body,
+      "persistentConnection": persistentConnection,
+      "bodyBytes": bodyBytes,
+      "reasonPhrase": reasonPhrase,
+    });
+  }
 }