+++ lib/src/utils/api_link.dart
@@ -1,14 +1,18 @@
 part of http_api;
 
-/// ApiLink that is mixed witch ChangeNotifier
-abstract class NotifierApiLink extends ApiLink with ChangeNotifier {}
-
 /// An abstract class that lets you create your own link
 ///
 /// Each link should extend this method
 abstract class ApiLink {
-  /// [ApiLink]s keeps reference to the first [ApiLink] in chain to simplify chaining
-  ApiLink _firstLink;
+  /// [ApiLink]s keeps reference to the first [ApiLink] in chain
+  /// to simplify chaining.
+  ///
+  /// Initialy set to this or null in release builds if `this` is a [DebugLink].
+  /// Can be changed by [chain] method.
+  ApiLink get _firstLink =>
+      __firstLink ?? (isReleaseBuild && (this is DebugLink) ? null : this);
+  @protected
+  ApiLink __firstLink;
   ApiLink _nextLink;
   bool _disposed = false;
   bool get disposed => false;
@@ -17,8 +21,8 @@ abstract class ApiLink {
   bool get chained => _nextLink != null;
 
   /// calls [callback] for every link in chain
-  void _forEach(ValueChanged<ApiLink> callback) {
-    ApiLink lastLink = _firstLink ?? this;
+  void _forEach(void Function(ApiLink) callback) {
+    var lastLink = _firstLink ?? this;
     do {
       callback(lastLink);
       lastLink = lastLink._nextLink;
@@ -27,7 +31,7 @@ abstract class ApiLink {
 
   /// Returns first [ApiLink] that matches provided test function.
   ApiLink _firstWhere(bool test(ApiLink link)) {
-    ApiLink lastLink = _firstLink ?? this;
+    var lastLink = _firstLink ?? this;
     do {
       if (test(lastLink)) return lastLink;
       lastLink = lastLink._nextLink;
@@ -35,50 +39,63 @@ abstract class ApiLink {
     return null;
   }
 
-  /// Closes all links
+  /// Marks all ApiLinks in chain as closed by setting
+  /// `closed` property to true.
+  ///
+  /// Closed links cannot be chained.
   void _closeChain() {
-    /// If [_firstLink] is not set, it will be set with [this].
-    /// Close link chain.
-    _firstLink ??= this;
-    _forEach((ApiLink link) {
+    _forEach((link) {
       link._closed = true;
     });
   }
 
+  @visibleForTesting
+  bool get isReleaseBuild {
+    var release = true;
+    assert(!(release = false));
+    return release;
+  }
+
   /// Chain multiple links into one.
+  /// throws [ApiError] when error occurs
   @nonVirtual
   ApiLink chain(ApiLink nextLink) {
-    if (nextLink == null)
-      throw ApiException(
+    if (nextLink == null) {
+      throw ApiError(
         "Cannot chain link $runtimeType with ${nextLink.runtimeType}\n"
         "nextLink cannot be null",
       );
+    }
 
-    if (closed || nextLink.closed)
-      throw ApiException(
+    if (closed || nextLink.closed) {
+      throw ApiError(
         "Cannot chain link $runtimeType with ${nextLink.runtimeType}\n"
         "You cannot chain links after attaching to BaseApi",
       );
+    }
 
-    if (disposed || nextLink.disposed)
-      throw ApiException(
+    if (disposed || nextLink.disposed) {
+      throw ApiError(
         "Cannot chain link $runtimeType with ${nextLink.runtimeType}\n"
         "You cannot chain disposed links",
       );
+    }
 
-    if (this is HttpLink)
-      throw ApiException(
+    if (this is HttpLink) {
+      throw ApiError(
         "Cannot chain link $runtimeType with ${nextLink.runtimeType}\n"
         "Adding link after http link will take no effect",
       );
+    }
 
-    /// If there is no chain, start it with current
-    if (_firstLink == null) {
-      _firstLink = this;
+    /// Do not chain (skip) [DebugLink] in release build.
+    if (isReleaseBuild) {
+      if (this is DebugLink) return nextLink;
+      if (nextLink is DebugLink) return this;
     }
 
     /// set [_firstLink] reference in [nextLink]
-    nextLink._firstLink = _firstLink;
+    nextLink.__firstLink = _firstLink;
 
     /// set reference to next link
     _nextLink = nextLink;
@@ -87,8 +104,8 @@ abstract class ApiLink {
   }
 
   /// This method is called on every request.
-  /// Calling `super.next` will cause invocation of `next` method in the next ApiLink
-  /// in the chain (if present)
+  /// Calling `super.next` will cause invocation of `next` method
+  /// in the next ApiLink in the chain (if present)
   @protected
   Future<ApiResponse> next(ApiRequest request) {
     return _nextLink?.next(request);
@@ -105,7 +122,4 @@ abstract class ApiLink {
     assert(_disposed == false, "ApiLink cannot be disposed more than once");
     _disposed = true;
   }
-
-  @override
-  String toString() => "ApiLink: ${this.runtimeType}";
 }