+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## [0.7.1] - 07 September 2020.
+- Removed flutter from dependency (add crossplatform support).
+  - Replaced Flutter `Key` class with `CacheKey`.
+- Cache system changes:
+  - `CacheKey` class now operates on `String` values.
+  - Create `CacheManager` class.
+  - Added `Cache` mixin that adds cache to `BaseApi` instances.
+    ```dart
+    class Api extends BaseApi with Cache {
+
+      @override
+      CacheManager createCacheManager() => InMemoryCache();
+
+      /// ** your custom Api class implementation **
+    }
+    ```
+- Now displaying `hexString` for id's in logger link.
+- Added `createdAt` property to `ApiRequest`. 
+- Made `saveCache` function optionally asynchronous.
+- url is now a required positional argument in BaseApi class constructor.
+- Improved readme file.
+- Bug fixes:
+  - Fixed example app compilation error.
+  - Disallowed chaining DebugLink in release apps when only one link is provided.
+## [0.6.0] - 06 September 2020.
+- Renamed current `DebugLink` class to `LoggerLink`.
+- Added abstract `DebugLink` class. DebugLinks are special types of links that will never be chained (will be skipped) in release builds. 
+- Replaced internal `ApiException`s with `ApiError`s.
 ## [0.5.0] - 05 September 2020.
 * Added cache support.
 * Added id property to ApiRequest and ApiResponse objects.